chatgpt/requirements.txt
django==4.1.7
djangorestframework==3.14.0
openai==0.26.5
django_redis==5.2.0
PyMySQL==1.0.2